+++ src/schemas/index.ts
@@ -7,6 +7,7 @@ import {
   listReviewSchema
 } from './list'
 import {
+  userCodeSchema,
   userIdSchema,
   userNotifySchema,
   userSetCommitteeMembersSchema,
@@ -20,6 +21,7 @@ export {
   listIdSchema,
   listOwnerSchema,
   listReviewSchema,
+  userCodeSchema,
   userIdSchema,
   userNotifySchema,
   userSetCommitteeMembersSchema,